JARVIS/requirements.txt
beautifulsoup4==4.12.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.3
colorama==0.4.6
comtypes==1.1.14
distlib==0.3.6
dlib @ file:///N:/my_stuff/Projects/Python/Mini_Project_Batch_B2/Mini_Project_Batch_B2/Transport_Administration_System_with_a_Voice_assistant_and_Drowsiness_detection/JARVIS/face_recog_dlib_file/dlib-19.19.0-cp37-cp37m-win_amd64.whl
downloadimages==0.0.1
face-recognition==1.3.0
face-recognition-models==0.3.0
filelock==3.10.6
googlesearch-python==1.2.3
idna==3.4
importlib-metadata==6.1.0
imutils==0.5.4
numpy==1.21.2
opencv-python==4.7.0.72
Pillow==9.5.0
platformdirs==3.2.0
playsound==1.2.2
pydub==0.25.1
pypiwin32==223
pyttsx3==2.90
pywin32==306
requests==2.28.2
scipy==1.7.3
soupsieve==2.4
SpeechRecognition==3.9.0
typing_extensions==4.5.0
urllib3==1.26.15
virtualenv==20.21.0
wikipedia==1.4.0
zipp==3.15.0
